Improvement for unconfirmed txs
There could be cases where the tx was included on a block, but not yet available on the SQL database. This could cause unconfirmed tx to go no "not found" and then to confirmed, being confusing for the user. For this reason, if the TX is not found, I do a last check in /tpool/confirmed: if it shows up here, but not in my SQL database I return the tx as unconfirmed. On the next check, once included on the SQL database it will show finally as confirmed